# gen/http/openapi.yaml
swagger: "2.0"
info:
    title: Film Library
    description: BiS Service for querying and managing films and actors info
    version: "1.0"
host: localhost:3239
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
...
    /films:
        get:
            tags:
                - SearchService
            summary: getAllFilms SearchService
            description: |4-
                **Required security scopes for jwt**:
                  * `api:read`
            operationId: SearchService#getAllFilms
            parameters:
                - name: Authorization
                  in: header
                  description: JWT used for authentication
                  required: false
                  type: string
                - name: GetAllFilmsRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/SearchServiceGetAllFilmsRequestBody'
                    required:
                        - SortBy
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/SearchServiceFilmResultResponseCollection'
                "401":
                    description: Unauthorized response.
                    schema:
                        type: string
                "403":
                    description: Forbidden response.
                    schema:
                        type: string
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
        post:
            tags:
                - FilmService
            summary: addFilm FilmService
            description: |4-
                **Required security scopes for jwt**:
                  * `api:write`
            operationId: FilmService#addFilm
            parameters:
                - name: Authorization
                  in: header
                  description: JWT used for authentication
                  required: false
                  type: string
                - name: AddFilmRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/FilmServiceAddFilmRequestBody'
                    required:
                        - FilmInfo
            responses:
                "201":
                    description: Created response.
                    schema:
                        $ref: '#/definitions/FilmServiceAddFilmResponseBody'
                "400":
                    description: Bad Request response.
                    schema:
                        $ref: '#/definitions/FilmServiceAddFilmAlreadyExistsResponseBody'
                        required:
                            - message
                            - id
                "401":
                    description: Unauthorized response.
                    schema:
                        type: string
                "403":
                    description: Forbidden response.
                    schema:
                        type: string
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
    /films/{FilmID}:
        put:
            tags:
                - FilmService
            summary: updateFilmInfo FilmService
            description: |4-
                **Required security scopes for jwt**:
                  * `api:write`
            operationId: FilmService#updateFilmInfo
            parameters:
                - name: FilmID
                  in: path
                  required: true
                  type: integer
                - name: Authorization
                  in: header
                  description: JWT used for authentication
                  required: false
                  type: string
                - name: UpdateFilmInfoRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/FilmServiceUpdateFilmInfoRequestBody'
                    required:
                        - FilmInfo
            responses:
                "201":
                    description: Created response.
                "401":
                    description: Unauthorized response.
                    schema:
                        type: string
                "403":
                    description: Forbidden response.
                    schema:
                        type: string
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
        delete:
            tags:
                - FilmService
            summary: deleteFilm FilmService
            description: |4-
                **Required security scopes for jwt**:
                  * `api:write`
            operationId: FilmService#deleteFilm
            parameters:
                - name: FilmID
                  in: path
                  description: Film ID
                  required: true
                  type: integer
                - name: Authorization
                  in: header
                  description: JWT used for authentication
                  required: false
                  type: string
            responses:
                "204":
                    description: No Content response.
                "401":
                    description: Unauthorized response.
                    schema:
                        type: string
                "403":
                    description: Forbidden response.
                    schema:
                        type: string
            schemes:
                - http
            security:
                - jwt_header_Authorization: []
...
definitions:
...
    FilmInfoRequestBody:
        title: FilmInfoRequestBody
        type: object
        properties:
            Actors:
                type: array
                items:
                    type: integer
                    example: 9529624566924527815
                    format: int64
                description: Actors' Ids
                example:
                    - 17093840019234325971
                    - 6600348201439270407
            Description:
                type: string
                description: Description of a film
                example: sqk
                maxLength: 1000
            Rating:
                type: number
                description: Rating (0.0 - 10.0)
                example: 2.5155373
                format: float
                minimum: 0
                maximum: 10
            ReleaseDate:
                type: string
                description: YYYY-MM-DD
                example: "2024-03-18"
                pattern: ^\d{4}\-(0[1-9]|1[012])\-(0[1-9]|[12][0-9]|3[01])$
            Title:
                type: string
                description: Title of a film
                example: 1n
                minLength: 1
                maxLength: 150
        description: Describes a Film to be added
        example:
            Actors:
                - 3064200069962830303
                - 904133416449501210
                - 14229838625793017532
            Description: i8m
            Rating: 1.2239147
            ReleaseDate: "2024-03-18"
            Title: "0"
        required:
            - Title
            - Description
            - ReleaseDate
            - Rating
            - Actors
...
    FilmResultResponse:
        title: 'Mediatype identifier: application/vnd.film; view=default'
        type: object
        properties:
            Actors:
                type: string
                example: Reiciendis et molestiae inventore molestias assumenda.
            Description:
                type: string
                example: Sint ea quidem ex.
            FilmID:
                type: integer
                description: Unique ID of a Film
                example: 239
                format: int64
            Rating:
                type: string
                example: Sapiente deleniti quae voluptas iure.
            ReleaseDate:
                type: string
                example: Quidem amet quaerat.
            Title:
                type: string
                example: Dolores earum amet ipsum rerum quia.
        description: FilmResultResponse result type (default view)
        example:
            Actors: Laboriosam voluptatem repellendus cum.
            Description: Sunt perferendis culpa.
            FilmID: 239
            Rating: Similique voluptas vel vel pariatur distinctio.
            ReleaseDate: Quo natus quam officiis aut.
            Title: Est tempora et et assumenda vel.
        required:
            - FilmID
    FilmServiceAddFilmAlreadyExistsResponseBody:
        title: FilmServiceAddFilmAlreadyExistsResponseBody
        type: object
        properties:
            id:
                type: string
                description: ID of existing data
                example: Vel molestiae.
            message:
                type: string
                description: Error message
                example: Entity already exists
        description: Film already exists
        example:
            id: Ullam repudiandae.
            message: Entity already exists
        required:
            - message
            - id
    FilmServiceAddFilmRequestBody:
        title: FilmServiceAddFilmRequestBody
        type: object
        properties:
            FilmInfo:
                $ref: '#/definitions/FilmInfoRequestBody'
        example:
            FilmInfo:
                Actors:
                    - 15409686236987074460
                    - 1963945278260898944
                Description: c6v
                Rating: 3.8577452
                ReleaseDate: "2024-03-18"
                Title: y2u
        required:
            - FilmInfo
    FilmServiceAddFilmResponseBody:
        title: 'Mediatype identifier: application/vnd.film; view=default'
        type: object
        properties:
            Actors:
                type: string
                example: In nulla iure.
            Description:
                type: string
                example: Sint laborum velit incidunt inventore.
            FilmID:
                type: integer
                description: Unique ID of a Film
                example: 239
                format: int64
            Rating:
                type: string
                example: Aut laboriosam fugiat consequuntur ab vero rem.
            ReleaseDate:
                type: string
                example: Eos asperiores consequatur ipsam.
            Title:
                type: string
                example: Sapiente qui eos impedit iste.
        description: AddFilmResponseBody result type (default view)
        example:
            Actors: Voluptates nesciunt.
            Description: Qui id voluptas quas earum.
            FilmID: 239
            Rating: Voluptas laudantium debitis expedita voluptas rerum.
            ReleaseDate: Assumenda enim sit.
            Title: Dolores ipsa odio.
        required:
            - FilmID
    FilmServiceUpdateFilmInfoRequestBody:
        title: FilmServiceUpdateFilmInfoRequestBody
        type: object
        properties:
            FilmInfo:
                $ref: '#/definitions/FilmInfoRequestBody'
        example:
            FilmInfo:
                Actors:
                    - 15409686236987074460
                    - 1963945278260898944
                Description: c6v
                Rating: 3.8577452
                ReleaseDate: "2024-03-18"
                Title: y2u
        required:
            - FilmInfo
...
    SearchServiceFilmResultResponseCollection:
        title: 'Mediatype identifier: application/vnd.film; type=collection; view=default'
        type: array
        items:
            $ref: '#/definitions/FilmResultResponse'
        description: GetAllFilmsResponseBody is the result type for an array of FilmResultResponse (default view)
        example:
            - Actors: Commodi provident fuga veritatis et quia neque.
              Description: Sapiente officiis rerum laudantium cupiditate.
              FilmID: 239
              Rating: Facere sit eos quod nulla at aspernatur.
              ReleaseDate: Perferendis aut quis eveniet quas eligendi animi.
              Title: Dolores sed architecto est et est perferendis.
            - Actors: Commodi provident fuga veritatis et quia neque.
              Description: Sapiente officiis rerum laudantium cupiditate.
              FilmID: 239
              Rating: Facere sit eos quod nulla at aspernatur.
              ReleaseDate: Perferendis aut quis eveniet quas eligendi animi.
              Title: Dolores sed architecto est et est perferendis.
            - Actors: Commodi provident fuga veritatis et quia neque.
              Description: Sapiente officiis rerum laudantium cupiditate.
              FilmID: 239
              Rating: Facere sit eos quod nulla at aspernatur.
              ReleaseDate: Perferendis aut quis eveniet quas eligendi animi.
              Title: Dolores sed architecto est et est perferendis.
    SearchServiceGetAllFilmsRequestBody:
        title: SearchServiceGetAllFilmsRequestBody
        type: object
        properties:
            SortBy:
                $ref: '#/definitions/SortByRequestBody'
        example:
            SortBy:
                Field: Rating
                Ordering: Descending
        required:
            - SortBy
...
    SortByRequestBody:
        title: SortByRequestBody
        type: object
        properties:
            Field:
                type: string
                description: Field to sort by (Rating (default) | Title | Release Date)
                default: Rating
                example: Rating
            Ordering:
                type: string
                description: Ascending / Descending
                default: Descending
                example: Descending
        description: Sorting configuration
        example:
            Field: Rating
            Ordering: Descending
        required:
            - Field
            - Ordering
securityDefinitions:
    basic_header_Authorization:
        type: basic
        description: |-
            Basic authentication used to authenticate security principal during sign-in

            **Security Scopes**:
              * `api:read`: Read-only access
    jwt_header_Authorization:
        type: apiKey
        description: |-
            Secures endpoint by requiring a valid JWT token retrieved via the signin endpoint.

            **Security Scopes**:
              * `api:read`: Read-only access
              * `api:write`: Read and write access
        name: Authorization
        in: header